third_party/WebKit/LayoutTests/accessibility/computed-name-expected.txt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS name is "alert name"
PASS name is "alertdialog name"
PASS name is "application name"
PASS name is "article name"
PASS implicitRoleElement.computedName is "implicit article name"
PASS name is "banner name"
PASS implicitRoleElement.computedName is "implicit banner name"
PASS name is "button name"
PASS implicitRoleElement.computedName is "implicit button name"
PASS implicitRoleElement.computedName is "implicit button name"
PASS implicitRoleElement.computedName is "implicit button name"
PASS name is "button name"
PASS implicitRoleElement.computedName is "implicit complementary name"
PASS name is "complementary name"
PASS name is "checkbox name"
PASS implicitRoleElement.computedName is "implicit checkbox name"
PASS name is "checkbox name"
PASS name is "dialog name"
PASS name is "directory name"
PASS name is "grid name"
PASS name is "row name"
PASS name is "columnheader name"
PASS name is "row name"
PASS name is "gridcell name"
PASS name is "combobox name"
PASS name is "definition name"
PASS name is "document name"
PASS implicitRoleElement.computedName is "implicit document name"
PASS name is "form name"
PASS name is "group name"
PASS implicitRoleElement.computedName is "implicit group name"
PASS implicitRoleElement.computedName is "implicit group name"
PASS name is "heading name"
PASS implicitRoleElement.computedName is "implicit heading name"
PASS implicitRoleElement.computedName is "implicit heading name"
PASS implicitRoleElement.computedName is "implicit heading name"
PASS implicitRoleElement.computedName is "implicit heading name"
PASS implicitRoleElement.computedName is "implicit heading name"
PASS implicitRoleElement.computedName is "implicit heading name"
PASS name is "heading name"
PASS name is "img name"
PASS implicitRoleElement.computedName is "implicit img name"
PASS name is "link name"
PASS implicitRoleElement.computedName is "implicit link name"
PASS implicitRoleElement.computedName is "implicit link name"
PASS name is "link name"
PASS name is "list name"
PASS implicitRoleElement.computedName is "implicit list name"
PASS implicitRoleElement.computedName is "implicit list name"
PASS name is "listbox name"
PASS implicitRoleElement.computedName is "implicit listbox name"
PASS name is "log name"
PASS name is "main name"
PASS implicitRoleElement.computedName is "implicit main name"
PASS name is "marquee name"
PASS name is "math name"
PASS name is "menu name"
PASS name is "menubar name"
PASS name is "menuitem name"
PASS name is "menuitem name"
PASS name is "menuitemcheckbox name"
PASS name is "menuitemcheckbox name"
PASS name is "menuitemradio name"
PASS name is "menuitemradio name"
PASS name is "navigation name"
PASS implicitRoleElement.computedName is "implicit navigation name"
PASS name is "note name"
PASS name is "option name"
PASS name is "option name"
PASS name is "progressbar name"
PASS implicitRoleElement.computedName is "implicit progressbar name"
PASS name is "radio name"
PASS implicitRoleElement.computedName is "implicit radio name"
PASS name is "radio name"
PASS name is "radiogroup name"
PASS name is "region name"
PASS implicitRoleElement.computedName is "implicit region name"
PASS name is "scrollbar name"
PASS name is "search name"
PASS name is "separator name"
PASS implicitRoleElement.computedName is "implicit separator name"
PASS name is "slider name"
PASS name is "spinbutton name"
PASS name is "status name"
PASS implicitRoleElement.computedName is "implicit status name"
PASS name is "tablist name"
PASS name is "tab name"
PASS name is "tab name"
PASS name is "tabpanel name"
PASS name is "textbox name"
PASS implicitRoleElement.computedName is "implicit textbox name"
PASS implicitRoleElement.computedName is "implicit textbox name"
PASS implicitRoleElement.computedName is "implicit textbox name"
PASS implicitRoleElement.computedName is "implicit textbox name"
PASS implicitRoleElement.computedName is "implicit textbox name"
PASS implicitRoleElement.computedName is "implicit textbox name"
PASS name is "timer name"
PASS name is "toolbar name"
PASS implicitRoleElement.computedName is "implicit toolbar name"
PASS name is "tooltip name"
PASS name is "tree name"
PASS name is "treegrid name"
PASS name is "treeitem name"
PASS name is "link name"
PASS svgElement.computedName is "Green rectangle"
PASS svgAElement.computedName is "link name"
PASS input.computedName is "Title0"
PASS input.computedName is "Label1"
PASS input.computedName is "AriaLabel2"
PASS input.computedName is "LabelledBy3"
PASS input.computedName is "Placeholder4"
PASS input.computedName is "Placeholder5"
PASS input.computedName is "LabelledBy6"
PASS input.computedName is "ARIA Placeholder8"
PASS input.computedName is "Placeholder9"
PASS successfullyParsed is true

TEST COMPLETE
